Create AuthorIcon component
