+++ components/AuthorIcon/types.ts
@@ -1,0 +1,3 @@
+export interface Author {
+  author: string | string[];
+}
